+++ provider/pkg/await/await.go
@@ -17,7 +17,6 @@ package await
 import (
 	"context"
 	"fmt"
-
 	"github.com/pulumi/pulumi-kubernetes/provider/v3/pkg/clients"
 	"github.com/pulumi/pulumi-kubernetes/provider/v3/pkg/cluster"
 	"github.com/pulumi/pulumi-kubernetes/provider/v3/pkg/kinds"