+++ tests/sdk/go/kustomize/main.go
@@ -10,10 +10,16 @@ import (
 func main() {
 	pulumi.Run(func(ctx *pulumi.Context) error {
 		ns, err := corev1.NewNamespace(ctx, "test", &corev1.NamespaceArgs{})
+		if err != nil {
+			return err
+		}
 		provider, err := k8s.NewProvider(ctx, "k8s", &k8s.ProviderArgs{
 			Kubeconfig: pulumi.String("~/.kube/config"),
 			Namespace:  ns.Metadata.Name(),
 		})
+		if err != nil {
+			return err
+		}
 
 		_, err = kustomize.NewDirectory(ctx, "helloWorld",
 			kustomize.DirectoryArgs{Directory: pulumi.String("helloWorld")},