+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Handle different namespaces for server-side diff (https://github.com/pulumi/pulumi-kubernetes/pull/1631)
 - Handle auto-named namespaces for server-side diff (https://github.com/pulumi/pulumi-kubernetes/pull/1633)
 - *Revert* Fix hanging updates for deployment await logic (https://github.com/pulumi/pulumi-kubernetes/pull/1596)
+- Use shared informer for await logic for deployments (https://github.com/pulumi/pulumi-kubernetes/pull/1639)
 
 ## 3.4.1 (June 24, 2021)